requirements.txt
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
Pillow==6.1.0
pip==20.0.2
pyTelegramBotAPI==3.6.7
request==2018.11.20
requests==2.22.0
TelegramBotAPI==0.3.2
wheel==0.33.4
